Log to stderr by default, not stdout
Go's log package writes to stderr by default (like all/most loggers).
However pion's logger by default writes to stdout by default, even
though pion's packages that use the logger override this to stderr.
So projects using both will end up with a mix of log messages to stdout
and stderr, which isn't desirable when trying to separate out the
log output. The unit test for the pion logger also writes to stdout,
which is inconsistent with pion's actual usage whereby the default
output is overridden with stderr. This PR changes the default
logging fd to stderr, and also changes the logger test to be
consistent with pion's usage in writing to stderr.